One JS page file per route
