+++ apps/block_scout_web/assets/js/pages/blocks.js
